[WIP] Move IntBlockPool slices to MemoryIndex
