+++ lucene/memory/src/java/org/apache/lucene/index/memory/MemoryIndex.java
@@ -57,11 +57,13 @@ import org.apache.lucene.util.BytesRefHash;
 import org.apache.lucene.util.BytesRefHash.DirectBytesStartArray;
 import org.apache.lucene.util.Counter;
 import org.apache.lucene.util.IntBlockPool;
-import org.apache.lucene.util.IntBlockPool.SliceReader;
-import org.apache.lucene.util.IntBlockPool.SliceWriter;
 import org.apache.lucene.util.RecyclingByteBlockAllocator;
 import org.apache.lucene.util.RecyclingIntBlockAllocator;
 import org.apache.lucene.util.Version;
+
+import static org.apache.lucene.util.IntBlockPool.INT_BLOCK_MASK;
+import static org.apache.lucene.util.IntBlockPool.INT_BLOCK_SHIFT;
+import static org.apache.lucene.util.IntBlockPool.INT_BLOCK_SIZE;
 
 /**
  * High-performance single-document main memory Apache Lucene fulltext search index.
@@ -164,6 +166,215 @@ import org.apache.lucene.util.Version;
  * </a>).
  */
 public class MemoryIndex {
+  private class SlicedIntBlockPool {
+    private IntBlockPool pool;
+    private SliceWriter writer;
+    private SliceReader reader;
+
+    /**
+     * An array holding the offset into the {@link IntBlockPool#LEVEL_SIZE_ARRAY} to quickly navigate
+     * to the next slice level.
+     */
+    private static final int[] NEXT_LEVEL_ARRAY = {1, 2, 3, 4, 5, 6, 7, 8, 9, 9};
+
+    /** An array holding the level sizes for int slices. */
+    private static final int[] LEVEL_SIZE_ARRAY = {2, 4, 8, 16, 16, 32, 32, 64, 64, 128};
+
+    /** The first level size for new slices */
+    private static final int FIRST_LEVEL_SIZE = LEVEL_SIZE_ARRAY[0];
+
+    private SlicedIntBlockPool(IntBlockPool pool) {
+      this.pool = pool;
+      writer = new SliceWriter(pool);
+      reader = new SliceReader(pool);
+    }
+
+    /**
+     * Creates a new int slice with the given starting size and returns the slices offset in the pool.
+     *
+     * @see SliceReader
+     */
+    private int newSlice(final int size) {
+      if (pool.intUpto > INT_BLOCK_SIZE - size) {
+        pool.nextBuffer();
+        assert assertSliceBuffer(pool.buffer);
+      }
+
+      final int upto = pool.intUpto;
+      pool.intUpto += size;
+      pool.buffer[pool.intUpto - 1] = 16;
+      return upto;
+    }
+
+    private static boolean assertSliceBuffer(int[] buffer) {
+      int count = 0;
+      for (int i = 0; i < buffer.length; i++) {
+        count += buffer[i]; // for slices the buffer must only have 0 values
+      }
+      return count == 0;
+    }
+
+    /** Allocates a new slice from the given offset */
+    private int allocSlice(final int[] slice, final int sliceOffset) {
+      final int level = slice[sliceOffset] & 15;
+      final int newLevel = NEXT_LEVEL_ARRAY[level];
+      final int newSize = LEVEL_SIZE_ARRAY[newLevel];
+      // Maybe allocate another block
+      if (pool.intUpto > INT_BLOCK_SIZE - newSize) {
+        pool.nextBuffer();
+        assert assertSliceBuffer(pool.buffer);
+      }
+
+      final int newUpto = pool.intUpto;
+      final int offset = newUpto + pool.intOffset;
+      pool.intUpto += newSize;
+      // Write forwarding address at end of last slice:
+      slice[sliceOffset] = offset;
+
+      // Write new level:
+      pool.buffer[pool.intUpto - 1] = 16 | newLevel;
+
+      return newUpto;
+    }
+
+    /**
+     * A {@link SliceWriter} that allows to write multiple integer slices into a given {@link
+     * IntBlockPool}.
+     *
+     * @see SliceReader
+     * @lucene.internal
+     */
+    private class SliceWriter {
+
+      private int offset;
+      private final IntBlockPool pool;
+
+      public SliceWriter(IntBlockPool pool) {
+        this.pool = pool;
+      }
+      /** */
+      public void reset(int sliceOffset) {
+        this.offset = sliceOffset;
+      }
+
+      /** Writes the given value into the slice and resizes the slice if needed */
+      public void writeInt(int value) {
+        int[] ints = pool.buffers[offset >> INT_BLOCK_SHIFT];
+        assert ints != null;
+        int relativeOffset = offset & INT_BLOCK_MASK;
+        if (ints[relativeOffset] != 0) {
+          // End of slice; allocate a new one
+          relativeOffset = allocSlice(ints, relativeOffset);
+          ints = pool.buffer;
+          offset = relativeOffset + pool.intOffset;
+        }
+        ints[relativeOffset] = value;
+        offset++;
+      }
+
+      /**
+       * starts a new slice and returns the start offset. The returned value should be used as the
+       * start offset to initialize a {@link SliceReader}.
+       */
+      public int startNewSlice() {
+        return offset = newSlice(FIRST_LEVEL_SIZE) + pool.intOffset;
+      }
+
+      /**
+       * Returns the offset of the currently written slice. The returned value should be used as the
+       * end offset to initialize a {@link SliceReader} once this slice is fully written or to reset
+       * the this writer if another slice needs to be written.
+       */
+      public int getCurrentOffset() {
+        return offset;
+      }
+    }
+
+    /**
+     * A {@link SliceReader} that can read int slices written by a {@link SliceWriter}
+     *
+     * @lucene.internal
+     */
+    private class SliceReader {
+
+      private final IntBlockPool pool;
+      private int upto;
+      private int bufferUpto;
+      private int bufferOffset;
+      private int[] buffer;
+      private int limit;
+      private int level;
+      private int end;
+
+      /** Creates a new {@link SliceReader} on the given pool */
+      public SliceReader(IntBlockPool pool) {
+        this.pool = pool;
+      }
+
+      /** Resets the reader to a slice give the slices absolute start and end offset in the pool */
+      public void reset(int startOffset, int endOffset) {
+        bufferUpto = startOffset / INT_BLOCK_SIZE;
+        bufferOffset = bufferUpto * INT_BLOCK_SIZE;
+        this.end = endOffset;
+        level = 0;
+
+        buffer = pool.buffers[bufferUpto];
+        upto = startOffset & INT_BLOCK_MASK;
+
+        final int firstSize = LEVEL_SIZE_ARRAY[0];
+        if (startOffset + firstSize >= endOffset) {
+          // There is only this one slice to read
+          limit = endOffset & INT_BLOCK_MASK;
+        } else {
+          limit = upto + firstSize - 1;
+        }
+      }
+
+      /**
+       * Returns <code>true</code> iff the current slice is fully read. If this method returns <code>
+       * true</code> {@link SliceReader#readInt()} should not be called again on this slice.
+       */
+      public boolean endOfSlice() {
+        assert upto + bufferOffset <= end;
+        return upto + bufferOffset == end;
+      }
+
+      /**
+       * Reads the next int from the current slice and returns it.
+       *
+       * @see SliceReader#endOfSlice()
+       */
+      public int readInt() {
+        assert !endOfSlice();
+        assert upto <= limit;
+        if (upto == limit) nextSlice();
+        return buffer[upto++];
+      }
+
+      private void nextSlice() {
+        // Skip to our next slice
+        final int nextIndex = buffer[limit];
+        level = NEXT_LEVEL_ARRAY[level];
+        final int newSize = LEVEL_SIZE_ARRAY[level];
+
+        bufferUpto = nextIndex / INT_BLOCK_SIZE;
+        bufferOffset = bufferUpto * INT_BLOCK_SIZE;
+
+        buffer = pool.buffers[bufferUpto];
+        upto = nextIndex & INT_BLOCK_MASK;
+
+        if (nextIndex + newSize >= end) {
+          // We are advancing to the final slice
+          assert end - nextIndex > 0;
+          limit = end - bufferOffset;
+        } else {
+          // This is not the final slice (subtract 4 for the
+          // forwarding address at the end of this new slice)
+          limit = upto + newSize - 1;
+        }
+      }
+    }
+  }
 
   private static final boolean DEBUG = false;
 
@@ -174,9 +385,8 @@ public class MemoryIndex {
   private final boolean storePayloads;
 
   private final ByteBlockPool byteBlockPool;
-  private final IntBlockPool intBlockPool;
-  //  private final IntBlockPool.SliceReader postingsReader;
-  private final IntBlockPool.SliceWriter postingsWriter;
+  private final SlicedIntBlockPool slicedIntBlockPool;
+  private final SlicedIntBlockPool.SliceWriter postingsWriter;
   private final BytesRefArray payloadsBytesRefs; // non null only when storePayloads
 
   private Counter bytesUsed;
@@ -237,19 +447,19 @@ public class MemoryIndex {
     final int maxBufferedIntBlocks =
         (int)
             ((maxReusedBytes - (maxBufferedByteBlocks * (long) ByteBlockPool.BYTE_BLOCK_SIZE))
-                / (IntBlockPool.INT_BLOCK_SIZE * (long) Integer.BYTES));
+                / (INT_BLOCK_SIZE * (long) Integer.BYTES));
     assert (maxBufferedByteBlocks * ByteBlockPool.BYTE_BLOCK_SIZE)
-            + (maxBufferedIntBlocks * IntBlockPool.INT_BLOCK_SIZE * Integer.BYTES)
+            + (maxBufferedIntBlocks * INT_BLOCK_SIZE * Integer.BYTES)
         <= maxReusedBytes;
     byteBlockPool =
         new ByteBlockPool(
             new RecyclingByteBlockAllocator(
                 ByteBlockPool.BYTE_BLOCK_SIZE, maxBufferedByteBlocks, bytesUsed));
-    intBlockPool =
+    slicedIntBlockPool = new SlicedIntBlockPool(
         new IntBlockPool(
             new RecyclingIntBlockAllocator(
-                IntBlockPool.INT_BLOCK_SIZE, maxBufferedIntBlocks, bytesUsed));
-    postingsWriter = new SliceWriter(intBlockPool);
+                INT_BLOCK_SIZE, maxBufferedIntBlocks, bytesUsed)));
+    postingsWriter = slicedIntBlockPool.writer;
     // TODO refactor BytesRefArray to allow us to apply maxReusedBytes option
     payloadsBytesRefs = storePayloads ? new BytesRefArray(bytesUsed) : null;
   }
@@ -842,7 +1052,7 @@ public class MemoryIndex {
       result.append(fieldName).append(":\n");
       SliceByteStartArray sliceArray = info.sliceArray;
       int numPositions = 0;
-      SliceReader postingsReader = new SliceReader(intBlockPool);
+      SlicedIntBlockPool.SliceReader postingsReader = slicedIntBlockPool.reader;
       for (int j = 0; j < info.terms.size(); j++) {
         int ord = info.sortedTerms[j];
         info.terms.get(ord, spare);
@@ -1652,7 +1862,7 @@ public class MemoryIndex {
 
     private class MemoryPostingsEnum extends PostingsEnum {
 
-      private final SliceReader sliceReader;
+      private final SlicedIntBlockPool.SliceReader sliceReader;
       private int posUpto; // for assert
       private boolean hasNext;
       private int doc = -1;
@@ -1663,7 +1873,7 @@ public class MemoryIndex {
       private final BytesRefBuilder payloadBuilder; // only non-null when storePayloads
 
       public MemoryPostingsEnum() {
-        this.sliceReader = new SliceReader(intBlockPool);
+        this.sliceReader = slicedIntBlockPool.reader;
         this.payloadBuilder = storePayloads ? new BytesRefBuilder() : null;
       }
 
@@ -1942,7 +2152,7 @@ public class MemoryIndex {
     fields.clear();
     this.normSimilarity = IndexSearcher.getDefaultSimilarity();
     byteBlockPool.reset(false, false); // no need to 0-fill the buffers
-    intBlockPool.reset(true, false); // here must must 0-fill since we use slices
+    slicedIntBlockPool.pool.reset(true, false); // here must must 0-fill since we use slices
     if (payloadsBytesRefs != null) {
       payloadsBytesRefs.clear();
     }
